refactor: extract dedupe and DB lifecycle from main.py
main.py shrinks from 371 to 206 lines (44% reduction) by extracting two
self-contained concerns:

modules/dedupe.py (143 lines)
  Pure business logic — no I/O, no scheduling, no external dependencies
  beyond datetime. Houses find_new_games, is_still_active, the grace-period
  helpers, and RECENTLY_EXPIRED_GRACE_PERIOD_HOURS. Easy to test in
  isolation; the existing tests in test_migrations.py keep working with
  patch targets pointing to the new module.

modules/db_lifecycle.py (55 lines)
  Database bootstrap helpers: run_db_migrations and verify_required_tables.
  Renamed from the old underscore-prefixed names to indicate they are now
  public, importable utilities rather than private main.py helpers.

main.py is now 99% orchestration: setup_logging, scheduler glue, and the
api-server-in-a-thread launcher (which stays here because it is tightly
coupled to the threading + scheduler setup, per the issue's explicit
out-of-scope note).

Tests updated:
  - patch("main.alembic_command.upgrade") → patch("modules.db_lifecycle.alembic_command.upgrade")
  - patch("main.psycopg2.connect")        → patch("modules.db_lifecycle.psycopg2.connect")
  - patch("main._run_db_migrations")      → patch("main.run_db_migrations")
  - patch("main._verify_required_tables") → patch("main.verify_required_tables")
  - main._find_new_games(...)             → main.find_new_games(...)

docs/architecture.md tree updated to list the new modules.

Closes #172
