+++ main.py
@@ -1,30 +1,23 @@
 import logging
-import os
 import threading
 import time
-from datetime import datetime, timedelta, timezone
 
-import psycopg2
 import requests
 import schedule
-from alembic.config import Config as AlembicConfig
 
-from alembic import command as alembic_command
 from config import (
     API_HOST,
     API_PORT,
     CHECK_INTERVAL_HOURS,
     DB_HOST,
-    DB_NAME,
-    DB_PASSWORD,
-    DB_PORT,
-    DB_USER,
     ENABLED_STORES,
     HEALTHCHECK_INTERVAL,
     SCHEDULE_TIME,
     TIMEZONE,
 )
 from modules.database import FreeGamesDatabase
+from modules.db_lifecycle import run_db_migrations, verify_required_tables
+from modules.dedupe import find_new_games, is_still_active
 from modules.healthcheck import healthcheck
 from modules.logging_config import setup_logging
 from modules.notifier import send_discord_message
@@ -46,131 +39,7 @@ class _HealthEndpointFilter(logging.Filter):
         return ' /health ' not in record.getMessage()
 
 
-def _is_still_active(game) -> bool:
-    """Return True if *game*'s promotion has not yet expired.
-
-    Games with an empty or un-parseable end_date are treated as still-active to
-    avoid false "new game" alerts caused by transient scraping failures.
-    """
-    end_date = game.end_date
-    if not end_date:
-        # Treat unknown end dates as active to avoid duplicate notifications.
-        return True
-
-    normalized = end_date.strip()
-    if normalized.endswith("Z"):
-        normalized = normalized[:-1] + "+00:00"
-
-    try:
-        ends_at = datetime.fromisoformat(normalized)
-    except ValueError:
-        # Keep legacy/malformed records from causing false "new" alerts.
-        return True
-
-    if ends_at.tzinfo is None:
-        ends_at = ends_at.replace(tzinfo=timezone.utc)
-
-    return ends_at >= datetime.now(timezone.utc)
-
-
-_RECENTLY_EXPIRED_GRACE_PERIOD_HOURS = 24
-
-
-def _recently_expired_urls(previous_games) -> set[str]:
-    """Return URLs whose promo expired within the last grace period.
-
-    Steam (and other stores) occasionally return a wrong end_date for a game
-    whose promo just ended (e.g. off-by-one year).  Without this guard,
-    a URL that expired minutes ago would pass both the ``previous_active_urls``
-    and ``previous_seen`` checks and trigger a duplicate notification.
-    """
-    now = datetime.now(timezone.utc)
-    grace = timedelta(hours=_RECENTLY_EXPIRED_GRACE_PERIOD_HOURS)
-    urls: set[str] = set()
-    for game in previous_games:
-        if not game.url or not game.end_date:
-            continue
-        normalized = game.end_date.strip()
-        if normalized.endswith("Z"):
-            normalized = normalized[:-1] + "+00:00"
-        try:
-            ends_at = datetime.fromisoformat(normalized)
-        except ValueError:
-            continue
-        if ends_at.tzinfo is None:
-            ends_at = ends_at.replace(tzinfo=timezone.utc)
-        if ends_at < now <= ends_at + grace:
-            urls.add(game.url)
-    return urls
-
-
-def _find_new_games(current_games, previous_games):
-    """Return games that are newly free compared to still-active previous promos.
-
-    Three checks prevent duplicate notifications:
-
-    1. ``previous_active_urls`` — URLs whose promos are still running.  A game
-       whose URL is already active is suppressed regardless of its end_date.
-    2. ``recently_expired_urls`` — URLs whose promo ended within the last
-       ``_RECENTLY_EXPIRED_GRACE_PERIOD_HOURS`` hours.  This prevents a store
-       returning a bad end_date (e.g. wrong year) right after expiry from
-       triggering a re-notification for the same promotion.
-    3. ``previous_seen`` — (url, end_date) pairs ever persisted.  Prevents
-       re-notification for an expired promo even if the URL no longer appears
-       in the active set.
-
-    A game that passes *all* checks is genuinely new or has started a fresh
-    promo with a different end_date after the grace period.
-
-    Same-run deduplication: ``notified_urls`` tracks URLs already added to
-    ``new_games`` in this loop so that a URL appearing twice in ``current_games``
-    (e.g. duplicate search-result rows) is only notified once.
-    """
-    # A (url, end_date) pair that already appeared in previous games should not
-    # trigger a new notification, regardless of whether the promo is still active.
-    # This prevents re-notifying for the same expired promo while still allowing
-    # re-notification when the same game has a new promo (different end_date).
-    previous_seen = {
-        (game.url, game.end_date)
-        for game in previous_games
-        if game.url
-    }
-
-    # Also track active URLs to suppress games seen before whose promos are still running.
-    previous_active_urls = {
-        game.url
-        for game in previous_games
-        if game.url and _is_still_active(game)
-    }
-
-    # Suppress re-notification for URLs whose promo just expired — store data
-    # errors (e.g. Steam returning a wrong year) would otherwise bypass both
-    # previous_active_urls and previous_seen when the end_date differs.
-    recently_expired = _recently_expired_urls(previous_games)
-
-    new_games = []
-    notified_urls: set[str] = set()
-    for game in current_games:
-        url = game.url
-        if url:
-            if (
-                url not in previous_active_urls
-                and url not in recently_expired
-                and (url, game.end_date) not in previous_seen
-                and url not in notified_urls
-            ):
-                new_games.append(game)
-                notified_urls.add(url)
-            continue
-
-        # Fallback for malformed records that do not have a url.
-        if game not in previous_games:
-            new_games.append(game)
-
-    return new_games
-
 def check_games():
-
     """Main execution function that checks for new free games and sends Discord notification."""
     logger.info("Checking for new free games...")
 
@@ -204,7 +73,7 @@ def check_games():
         logger.error("Failed to load previous games: %s", e)
         return
 
-    new_games = _find_new_games(current_games, previous_games)
+    new_games = find_new_games(current_games, previous_games)
 
     if new_games:
         logger.info("Found %d new free games! Sending Discord notification...", len(new_games))
@@ -251,7 +120,7 @@ def check_games():
     stores_with_results = {g.store for g in current_games}
     preserved = [
         g for g in previous_games
-        if g.store not in stores_with_results and _is_still_active(g)
+        if g.store not in stores_with_results and is_still_active(g)
     ]
     if preserved:
         logger.info(
@@ -271,41 +140,6 @@ def check_games():
     except Exception as e:
         logger.error("Unexpected error saving games: %s", e)
         logger.warning("Failed to update local cache.")
-
-def _run_db_migrations():
-    """Apply any pending Alembic migrations up to the latest revision."""
-    logger.info("Applying database migrations...")
-    # Suppress verbose per-revision Alembic log lines from service logs.
-    # env.py skips fileConfig when the service's logging is already configured,
-    # but raise the level here as well to guard against any propagation.
-    logging.getLogger("alembic").setLevel(logging.WARNING)
-    cfg = AlembicConfig(os.path.join(os.path.dirname(__file__), "alembic.ini"))
-    alembic_command.upgrade(cfg, "head")
-    logger.info("Database migrations applied successfully.")
-
-
-def _verify_required_tables():
-    """Fail fast when required DB tables are missing after migrations."""
-    logger.info("Verifying required database tables...")
-
-    conn_params = {
-        "host": DB_HOST,
-        "port": DB_PORT,
-        "dbname": DB_NAME,
-        "user": DB_USER,
-        "password": DB_PASSWORD,
-    }
-
-    with psycopg2.connect(**conn_params) as conn:
-        with conn.cursor() as cursor:
-            cursor.execute("SELECT to_regclass('free_games.last_notification')")
-            if cursor.fetchone()[0] is None:
-                raise RuntimeError(
-                    "Required table free_games.last_notification is missing after migrations. "
-                    "Run 'alembic upgrade head' and verify DB permissions."
-                )
-
-    logger.info("Required database tables verified successfully.")
 
 
 def _start_api_server():
@@ -327,8 +161,8 @@ def main():
         logger.info("Database configuration detected. Initializing database...")
         db = FreeGamesDatabase()
         db.init_db()
-        _run_db_migrations()
-        _verify_required_tables()
+        run_db_migrations()
+        verify_required_tables()
     else:
         logger.info("No database configuration detected. Using JSON file storage.")
 
@@ -365,6 +199,7 @@ def main():
     while True:
         schedule.run_pending()
         time.sleep(1)
+
 
 if __name__ == "__main__":
     logger.info("Starting service...")
